Changed est_p to int16_t, several other small changes.
Trying to track down why this program is consistently slightly
worse than v1, and on less proper simulated data it was worse
enough that we care. None of these changes resulted in major
improvements though.

Also, making the CRF smaller with RF the same makes the results
worse not better by either mean accuracy or variance. One of the
major points of this code is this feature which v1 could not
accomodate, but it is proving counter-intuitive in its effects. Do
I need a conditional dependence term due to overlapping RF?

+++ load-input.cpp
@@ -515,7 +515,7 @@ static void set_crf_points(input_t *input) {
       for(int i=0; i < input->n_windows; i++)
 	sample->msp[h][i] = sample->apriori_subpop;
 
-      MA(sample->est_p[h], sizeof(int8_t)*input->n_windows*n_subpops, int8_t);
+      MA(sample->est_p[h], sizeof(int16_t)*input->n_windows*n_subpops, int16_t);
     }
   }
   fprintf(stderr,"done\n");
